fix travis??

+++ default.nix
@@ -14,17 +14,27 @@ let
     echo >&2
     echo >&2 "Installing Nix-Darwin..."
     echo >&2
-    ${pkgs.lib.optionalString pkgs.stdenvNoCC.isDarwin ''
-      cd /etc
-      for file in bashrc shells skhdrc zprofile zshenv zshrc nix/nix.conf; do
-          # if an /etc config file isn't a symlink, then we should move it
-          [[ ! -L $file ]] && sudo mv $file "$file.bak" && echo "backed up $file"
-      done
 
-      # create the inital generation
-      $(nix-build ${sources.nix-darwin} -A system --no-out-link)/sw/bin/darwin-rebuild build
-      $(nix-build ${sources.nix-darwin} -A system --no-out-link)/sw/bin/darwin-rebuild switch --flake ${configuration}
-    ''}
+    # setup nix-darwin global store
+    sudo mkdir -p /run || sudo ln -s private/var/run /run
+
+    # back up
+    cd /etc
+    for file in bashrc shells skhdrc zprofile zshenv zshrc nix/nix.conf; do
+        # if an /etc config file isn't a symlink, then we should move it
+        [[ -e $file ]] && [[ ! -L $file ]] && sudo mv $file "$file.bak" && echo "backed up $file"
+    done
+
+    echo darwin-config=${sources.nix-darwin}/modules/examples/simple.nix
+    export NIX_PATH=darwin-config=${sources.nix-darwin}/modules/examples/simple.nix:darwin=${sources.nix-darwin}
+
+    # build nix darwin
+    nix-build ${sources.nix-darwin} -A system --no-out-link build
+
+    export NIX_PATH=darwin-config=${configuration}:darwin=${sources.nix-darwin}
+
+    # build the actual configuration once we do that
+    sudo ./result/sw/bin/darwin-rebuild switch --flake ".#randall"
   '';
 
   darwinRebuild = pkgs.writeShellScriptBin "rebuild" ''
